Changed rotation controller so that the rotate() method returns a double, shortened the tsop library, removing methods and variables that wearnt needed, added some methods to the Light library, including averaging the angle and returning the angle, made direction controller the sole controller of all movement. IE all movement (rotation or directional) now has to go through the direction controller library. It now calcs direction based on light, tsops, compass and pixy (kinda)

+++ Software/lib/RotationController/RotationController.cpp
@@ -37,7 +37,7 @@ void RotationController::calcRotation(){
     compass.updateTarget(newTarget);
 }
 
-int RotationController::rotate(){
+double RotationController::rotate(){
     finalRotation = compass.read();
     return finalRotation;
 }
